refactor compatibility
